move kerchunk parsers into their own internal module
