+++ virtualizarr/parsers/__init__.py
@@ -1,8 +1,8 @@
 from virtualizarr.parsers.dmrpp import DMRPPParser
 from virtualizarr.parsers.fits import FITSParser
 from virtualizarr.parsers.hdf import HDFParser
-from virtualizarr.parsers.kerchunk_json import KerchunkJSONParser
-from virtualizarr.parsers.kerchunk_parquet import KerchunkParquetParser
+from virtualizarr.parsers.kerchunk.json import KerchunkJSONParser
+from virtualizarr.parsers.kerchunk.parquet import KerchunkParquetParser
 from virtualizarr.parsers.netcdf3 import NetCDF3Parser
 from virtualizarr.parsers.zarr import ZarrParser
