+++ virtualizarr/parsers/kerchunk/parquet.py
